# server/openapi.yaml
openapi: 3.0.3
info:
  title: TinyIce API
  description: |
    REST API for TinyIce audio streaming server.

    ## Authentication
    Two authentication methods are supported:

    **Session cookie** — Log in via `POST /login` to get a `sid` session cookie.

    **Bearer token** — Create an API token in the admin panel (`/admin/tokens`) and pass it
    as `Authorization: Bearer ti_...` header. Tokens can have optional expiry dates and
    track last-used time/IP.

    No CSRF tokens are needed for JSON API requests — the `Content-Type: application/json`
    header provides implicit CSRF protection since it triggers a CORS preflight for
    cross-origin requests.

    ## Real-time Updates
    Use the SSE endpoint `GET /events` for real-time metadata, stream status, and
    listener count updates without polling.
  version: "2.0"
  license:
    name: MIT
servers:
  - url: /
    description: This server
...
paths:
...
  /api/streams:
    get:
      tags: [Streams]
      summary: List all streams
      description: |
        Returns all stream mounts the authenticated user has access to.
        Includes both active streams (with a connected source) and configured-but-offline mounts.
        Offline mounts have `live: false` and empty `source_ip`.
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: Array of stream info objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Streams]
      summary: Create a stream mount
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mount, password]
              properties:
                mount:
                  type: string
                  description: Mount path (will be prefixed with / if missing)
                  example: /live
                password:
                  type: string
                  description: Source password for this mount
                  example: secret123
                burst_size:
                  type: integer
                  description: Initial burst size in bytes for this mount
                  example: 65536
                max_listeners:
                  type: integer
                  description: Mount-specific listener cap (0 = use global cap)
                  example: 250
      responses:
        "200":
          description: Mount created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Mount already taken
    put:
      tags: [Streams]
      summary: Update stream mount settings
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mount]
              properties:
                mount:
                  type: string
                  description: Existing mount path
                  example: /live
                burst_size:
                  type: integer
                  description: Burst size in bytes for this mount
                  example: 131072
                max_listeners:
                  type: integer
                  description: Mount-specific listener cap (0 = use global cap)
                  example: 500
      responses:
        "200":
          description: Mount settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Streams]
      summary: Delete a stream mount
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mount deleted
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sid
      description: Session cookie obtained via POST /login
    bearer:
      type: http
      scheme: bearer
      description: API token created in the admin panel (format ti_...)
  responses:
    Unauthorized:
      description: Not authenticated — session cookie missing or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request body or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions (e.g. requires superadmin)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: created
        mount:
          type: string
    Stream:
      type: object
      properties:
        mount:
          type: string
        content_type:
          type: string
          description: MIME type (e.g. audio/mpeg)
        bitrate:
          type: string
          description: Bitrate string (e.g. "128")
        burst_size:
          type: integer
          description: Burst size in bytes for this mount
        listeners:
          type: integer
        max_listeners:
          type: integer
          description: Mount-specific listener cap (0 = use global cap)
        source_ip:
          type: string
          description: IP of connected source, empty if offline
        visible:
          type: boolean
          description: Whether shown on public pages
        enabled:
          type: boolean
        health:
          type: number
          description: Stream health percentage (0-100)
        uptime:
          type: string
          description: Human-readable uptime
        current_song:
          type: string
        name:
          type: string
        status:
          type: string
          example: dead
        status_class:
          type: string
          example: health_dead
        status_reason:
          type: string
          example: no data for 63s
        last_error:
          type: string
        status_updated_at:
          type: integer
          format: int64
        last_recovery_at:
          type: integer
          format: int64
        last_recovery_result:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticEntry"
        live:
          type: boolean
          description: Whether a source is currently connected
...
    DiagnosticEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        class:
          type: string
        reason:
          type: string
        error:
          type: string
        actor:
          type: string
...